+++ Eclipse/src/Graphics/Grid/AABBTree.h
@@ -46,7 +46,7 @@ namespace Eclipse
         std::vector<unsigned int> CheckOverlapAgainstGrid(AABBComponent& object);
         std::vector<unsigned int> CheckOverlap(DYN_AABB& object);
         std::vector<unsigned int> CheckMouseOverlapAgainstGrid(glm::vec3& rayStart, glm::vec3& rayDir, float& t);
-        unsigned int NumberOfIntersections(std::vector<unsigned int>& in);
+        size_t NumberOfIntersections(std::vector<unsigned int>& in);
 
     private:
         unsigned AllocateNode();
