+++ Eclipse/src/Graphics/ModelFactory/AssimpModel/AssimpManager.cpp
@@ -79,7 +79,7 @@ namespace Eclipse
         glUniform1i(NoTexures, 0);
     }
 
-    unsigned int AssimpModelManager::MeshFactoryCount()
+    size_t AssimpModelManager::MeshFactoryCount()
     {
         return AssimpLoadedModels.size();
     }
@@ -424,7 +424,7 @@ namespace Eclipse
         // See how many Models
         GeometryFileRead.read(reinterpret_cast<char*>(&TotalNumberOfModels), sizeof(TotalNumberOfModels));
 
-        for (int i = 0; i < TotalNumberOfModels; i++)
+        for (unsigned int i = 0; i < TotalNumberOfModels; i++)
         {
             VerticesSize = 0;
             IndicesSize = 0;
@@ -686,7 +686,7 @@ namespace Eclipse
         return AssimpModelContainer_;
     }
 
-    unsigned int AssimpModelManager::AssimpModelCount()
+    size_t AssimpModelManager::AssimpModelCount()
     {
         return AssimpModelContainer_.size();
     }
@@ -809,7 +809,7 @@ namespace Eclipse
         return AssimpModelContainerV2;
     }
 
-    unsigned int AssimpModelManager::MeshModelCount()
+    size_t AssimpModelManager::MeshModelCount()
     {
         return AssimpModelContainerV2.size();
     }
